+++ R/plot_mesh.R
@@ -24,7 +24,9 @@ plot_mesh <- function(mesh, spatial_data = NULL, longitude_column = "LONG", lati
     )
   }
 
-  spatial_mesh <- fdmr::mesh_to_spatial(mesh = mesh, crs = expected_crs)
+  spatial_mesh_original <- fdmr::mesh_to_spatial(mesh = mesh, crs = expected_crs)
+
+  spatial_mesh <- fdmr::antimeridian_wrapping(spatial_mesh_original, crs = expected_crs, unique_inst = FALSE, to_sp = FALSE)
 
   plot_polygons <- FALSE
   plot_points <- FALSE
@@ -60,6 +62,8 @@ plot_mesh <- function(mesh, spatial_data = NULL, longitude_column = "LONG", lati
   m <- leaflet::leaflet()
   m <- leaflet::addTiles(m, group = "OSM")
   m <- leaflet::addPolygons(m, data = spatial_mesh, weight = 0.5, fillOpacity = 0.2, fillColor = "#5252ca", group = "Mesh")
+  m <- leaflet::addMeasure(m, position = "bottomleft", primaryLengthUnit = 'kilometers', primaryAreaUnit = 'sqmeters')
+  m <- leafem::addMouseCoordinates(m, native.crs = TRUE)
 
   if (plot_polygons) {
     m <- leaflet::addPolygons(m, data = spatial_data, fillColor = "#d66363", color = "green", weight = 1, group = "Spatial")